Add pull request feedback and add a test
Only passing token (or nil) to the build_url function rather than all the options
Deleting the token option from the options hash correctly
Updated all the public functions to include an optional options argument that is passed then passed onto request()
Added a test to confirm custom token is actually used in the generated url.
Note: I've modified the test setup. Not sure why the token env var was only set if it was missing.
I feel like we can just override it within the test environment to be what we expect which makes the tests easier to write and reason about.