Add environment variable checks.
Instead of just using a given value, this commit adds the ability to use
a environment variable instead. It uses the following syntax in
`config.exs`:

        config :nadia, :token, {:system, "ENV_VAR_NAME", "default value"}

It is possible to define a default value as well. The old API still
works, it just adds a way to use env vars as well.